+++ waspc/data/Generator/templates/sdk/wasp/auth/session.ts
@@ -1,15 +1,15 @@
 {{={= =}=}}
 import { Request as ExpressRequest } from "express";
 
-import { type {= userEntityUpper =} } from "wasp/entities"
-import { type AuthUser } from 'wasp/auth'
+import { type {= userEntityUpper =} } from '../entities/index.js';
+import { type AuthUserData } from '../server/auth/user.js';
 
 import { auth } from "./lucia.js";
 import type { Session } from "lucia";
 import { throwInvalidCredentialsError } from "./utils.js";
 
 import { prisma } from 'wasp/server';
-import { createAuthUser } from "../server/auth/user.js";
+import { createAuthUserData } from "../server/auth/user.js";
 
 // PRIVATE API
 // Creates a new session for the `authId` in the database
@@ -17,52 +17,42 @@ export async function createSession(authId: string): Promise<Session> {
   return auth.createSession(authId, {});
 }
 
+type SessionAndUser = {
+  session: Session;
+  user: AuthUserData;
+}
+
 // PRIVATE API
-export async function getSessionAndUserFromBearerToken(req: ExpressRequest): Promise<{
-  user: AuthUser | null,
-  session: Session | null,
-}> {
+export async function getSessionAndUserFromBearerToken(req: ExpressRequest): Promise<SessionAndUser | null> {
   const authorizationHeader = req.headers["authorization"];
 
   if (typeof authorizationHeader !== "string") {
-    return {
-      user: null,
-      session: null,
-    };
+    return null;
   }
 
   const sessionId = auth.readBearerToken(authorizationHeader);
   if (!sessionId) {
-    return {
-      user: null,
-      session: null,
-    };
+    return null;
   }
 
   return getSessionAndUserFromSessionId(sessionId);
 }
 
 // PRIVATE API
-export async function getSessionAndUserFromSessionId(sessionId: string): Promise<{
-  user: AuthUser | null,
-  session: Session | null,
-}> {
+export async function getSessionAndUserFromSessionId(sessionId: string): Promise<SessionAndUser | null> {
   const { session, user: authEntity } = await auth.validateSession(sessionId);
 
   if (!session || !authEntity) {
-    return {
-      user: null,
-      session: null,
-    };
+    return null;
   }
 
   return {
     session,
-    user: await getUser(authEntity.userId)
+    user: await getAuthUserData(authEntity.userId)
   }
 }
 
-async function getUser(userId: {= userEntityUpper =}['id']): Promise<AuthUser> {
+async function getAuthUserData(userId: {= userEntityUpper =}['id']): Promise<AuthUserData> {
   const user = await prisma.{= userEntityLower =}
     .findUnique({
       where: { id: userId },
@@ -79,7 +69,7 @@ async function getUser(userId: {= userEntityUpper =}['id']): Promise<AuthUser> {
     throwInvalidCredentialsError()
   }
 
-  return createAuthUser(user);
+  return createAuthUserData(user);
 }
 
 // PRIVATE API