+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/server/auth/user.ts
@@ -8,27 +8,30 @@ import {
   deserializeAndSanitizeProviderData
 } from '../../auth/utils.js'
 import { type ProviderName } from '../_types/index.js'
-import { getFirstProviderUserId } from '../../auth/user.js'
 import { Expand } from '../../universal/types.js'
 
 // PUBLIC API
+export type AuthUser = AuthUserData & {
+  getFirstProviderUserId: () => string | null,
+}
+
+// PRIVATE API
 /**
  * Ideally, we'd do something like this:
  * ```
- * export type AuthUser = ReturnType<typeof createAuthUser>
+ * export type AuthUserData = ReturnType<typeof createAuthUserData>
  * ```
- * to get the benefits of the createAuthUser and the AuthUser type being in sync.
+ * to get the benefits of the createAuthUser and the AuthUserData type being in sync.
  * 
  * But since we are not using strict mode, the inferred return type of createAuthUser
- * is not correct. So we have to define the AuthUser type manually.
+ * is not correct. So we have to define the AuthUserData type manually.
  * 
  * TODO: Change this once/if we switch to strict mode. https://github.com/wasp-lang/wasp/issues/1938
  */
-export type AuthUser = Omit<UserEntityWithAuth, 'auth'> & {
+export type AuthUserData = Omit<UserEntityWithAuth, 'auth'> & {
   identities: {
     google: Expand<UserFacingProviderData<'google'>> | null
   },
-  getFirstProviderUserId: () => string | null,
 }
 
 type UserFacingProviderData<PN extends ProviderName> = {
@@ -46,7 +49,7 @@ export type AuthEntityWithIdentities = Auth & {
 }
 
 // PRIVATE API
-export function createAuthUser(user: UserEntityWithAuth): AuthUser {
+export function createAuthUserData(user: UserEntityWithAuth): AuthUserData {
   const { auth, ...rest } = user
   if (!auth) {
     throw new Error(`🐝 Error: trying to create a user without auth data.
@@ -58,7 +61,6 @@ This should never happen, but it did which means there is a bug in the code.`)
   return {
     ...rest,
     identities,
-    getFirstProviderUserId: () => getFirstProviderUserId(user),
   }
 }