+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/server/utils.ts
@@ -1,10 +1,10 @@
 import { Request, Response, NextFunction } from 'express'
 
-import { type AuthUser } from 'wasp/auth'
+import { type AuthUserData } from './auth/user.js'
 
 type RequestWithExtraFields = Request & {
-  user?: AuthUser;
-  sessionId?: string;
+  user: AuthUserData | null;
+  sessionId: string | null;
 }
 
 /**