+++ waspc/headless-test/examples/todoApp/src/client/pages/ProfilePage.tsx
@@ -2,7 +2,6 @@ import React, { useEffect } from 'react'
 import { type AuthUser as User } from 'wasp/auth'
 import { Link } from 'wasp/client/router'
 import { api } from 'wasp/client/api'
-import { getName } from '../user'
 
 async function fetchCustomRoute() {
   const res = await api.get('/foo/bar')
@@ -14,13 +13,11 @@ export const ProfilePage = ({ user }: { user: User }) => {
     fetchCustomRoute()
   }, [])
 
-  const name = getName(user)
-
   return (
     <>
       <h2>Profile page</h2>
       <div>
-        Hello <strong>{name}</strong>! Your status is{' '}
+        Hello <strong>{user.getFirstProviderUserId()}</strong>! Your status is{' '}
         <strong>
           {user.identities.email && user.identities.email.isEmailVerified
             ? 'verfied'